OCR/requirements.txt
Pillow
regex
